Modify factory method for revolute joints to take in all parameters
It seems as if the C++ and Rust representations of b2RevoluteJointDef are misaligned after the base class portion. The alignment is off by 4 bytes after the base class fields, which causes all the b2RevoluteJointDef fields after that to be incorrect. This solution brute forces it by removing the need for b2(Revolute)JointDef entirely.